refactor: fix package of JSONObject
* `com.ichi2.utils` => `com.ichi2.anki.common.utils.ext`
* file was in correct location, package didn't match

Looked like a circular dependency of libAnki

Issue 18015
