+++ AnkiDroid/src/main/java/com/ichi2/libanki/NotetypeJson.kt
@@ -22,9 +22,9 @@ import anki.notetypes.StockNotetype.OriginalStockKind.ORIGINAL_STOCK_KIND_IMAGE_
 import anki.notetypes.StockNotetype.OriginalStockKind.ORIGINAL_STOCK_KIND_UNKNOWN_VALUE
 import com.ichi2.anki.api.AddContentApi.Companion.DEFAULT_DECK_ID
 import com.ichi2.anki.common.json.NamedObject
+import com.ichi2.anki.common.utils.ext.deepClone
 import com.ichi2.anki.common.utils.ext.toStringList
 import com.ichi2.utils.JSONObjectHolder
-import com.ichi2.utils.deepClone
 import org.json.JSONArray
 import org.json.JSONException
 import org.json.JSONObject
