+++ common/src/main/java/com/ichi2/anki/common/utils/ext/JSONObject.kt
@@ -39,11 +39,10 @@
 *    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 *    SOFTWARE.
 */
-package com.ichi2.utils
+package com.ichi2.anki.common.utils.ext
 
 import androidx.annotation.CheckResult
 import androidx.annotation.VisibleForTesting
-import com.ichi2.anki.common.utils.ext.deepClone
 import org.json.JSONArray
 import org.json.JSONObject
